+++ src/main/java/com/gregtechceu/gtceu/api/item/component/ElectricStats.java
@@ -46,14 +46,14 @@ public class ElectricStats implements IInteractionItem, ISubItemHandler, IAddInf
     public final boolean chargeable;
     public final boolean dischargeable;
 
-    protected ElectricStats(long maxCharge, long tier, boolean chargeable, boolean dischargeable) {
+    protected ElectricStats(long maxCharge, int tier, boolean chargeable, boolean dischargeable) {
         this.maxCharge = maxCharge;
-        this.tier = (int) tier;
+        this.tier = tier;
         this.chargeable = chargeable;
         this.dischargeable = dischargeable;
     }
 
-    public static ElectricStats create(long maxCharge, long tier, boolean chargeable, boolean dischargeable) {
+    public static ElectricStats create(long maxCharge, int tier, boolean chargeable, boolean dischargeable) {
         return new ElectricStats(maxCharge, tier, chargeable, dischargeable);
     }
 
@@ -231,7 +231,7 @@ public class ElectricStats implements IInteractionItem, ISubItemHandler, IAddInf
         }
     }
 
-    public static ElectricStats createElectricItem(long maxCharge, long tier) {
+    public static ElectricStats createElectricItem(long maxCharge, int tier) {
         return ElectricStats.create(maxCharge, tier, true, false);
     }
